requirements.txt
Pillow==5.4.1
pyimgur==0.6.0
psutil==5.6.0
arrow==0.13.1
requests==2.21.0
argparse==1.4.0
